tests.utils: Create `AutoNowOverrideMixin()` from existing code
